backend/pyproject.toml
[project]
name = "music-download-backend"
version = "1.0.0"
description = "Music Download Backend - FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
    "qqmusic-api-python>=0.5.2",
    "httpx>=0.28.1",
]
